@@ -7,52 +7,56 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.0-262.2.beta <4.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_riverpod: ^2.3.6
-  riverpod_annotation: ^2.1.1
-  watch_connectivity: ^0.1.6
+  flutter_riverpod: ^2.6.1
+  riverpod_annotation: ^2.6.1
+  watch_connectivity: ^0.2.1+1
   flutter_ftms:
     # path: ./../flutter_ftms
     git:
       url: https://github.com/hifiaz/flutter_ftms.git
       ref: main 
-  firebase_core: ^2.15.0
-  firebase_auth: ^4.7.2
-  cloud_firestore: ^4.8.4
-  firebase_storage: ^11.2.5
+  firebase_core: ^4.0.0
+  firebase_auth: ^6.0.1
+  cloud_firestore: ^6.0.0
+  firebase_storage: ^13.0.0
   dartz: ^0.10.1
-  freezed_annotation: ^2.4.1
-  riverpod: ^2.3.6
+  freezed_annotation: ^3.1.0
+  riverpod: ^2.6.1
   chart_sparkline: ^1.0.13
-  health: ^7.0.1
-  go_router: ^10.0.0
+  health: ^13.1.3
+  go_router: ^16.2.0
   equatable: ^2.0.5
   logger: ^2.0.1
-  intl: ^0.18.1
-  liquid_progress_indicator:
-    git:
-     url: https://github.com/Jollastro/liquid_progress_indicator.git
+  intl: ^0.20.2
+
   stop_watch_timer: ^3.0.1
-  flutter_local_notifications: ^16.0.0-dev.1
-  firebase_messaging: ^14.6.5
+  flutter_local_notifications: ^19.4.1
+  firebase_messaging: ^16.0.0
+  google_sign_in: ^6.2.1
+  sign_in_with_apple: ^6.1.3
+  crypto: ^3.0.5
+  shared_preferences: ^2.2.2
+
+  flutter_launcher_icons: ^0.14.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
-  riverpod_generator: ^2.2.4
-  build_runner: ^2.4.6
-  custom_lint: ^0.4.0
-  riverpod_lint: ^1.4.0
-  freezed: ^2.4.1
-  json_serializable: ^6.7.1
+  flutter_lints: ^6.0.0
+  riverpod_generator: ^2.6.5
+  build_runner: ^2.4.13
+  custom_lint: ^0.7.0
+  riverpod_lint: ^2.6.5
+  freezed: ^3.1.0
+  json_serializable: ^6.8.0
 
 flutter:
   uses-material-design: true
