@@ -5,13 +5,15 @@ import 'package:flutter/material.dart';
 class ItemFasting extends StatelessWidget {
   final FastingModel fasting;
   final IconData icon;
-  final Color color;
+  final Color? color;
+  final Gradient? gradient;
   const ItemFasting({
-    Key? key,
+    super.key,
     required this.fasting,
     required this.icon,
-    required this.color,
-  }) : super(key: key);
+    this.color,
+    this.gradient,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -25,28 +27,64 @@ class ItemFasting extends StatelessWidget {
       },
       child: Container(
         width: double.infinity,
-        margin: const EdgeInsets.symmetric(horizontal: 16.0),
-        padding: const EdgeInsets.all(16.0),
+        padding: const EdgeInsets.all(24.0),
         decoration: BoxDecoration(
-            color: color,
-            borderRadius: const BorderRadius.all(Radius.circular(10))),
+          color: color,
+          gradient: gradient,
+          borderRadius: BorderRadius.circular(20),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.05),
+              blurRadius: 10,
+              offset: const Offset(0, 4),
+            ),
+          ],
+        ),
         child: Row(
           children: [
             Expanded(
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  Text(fasting.title,
-                      style: Theme.of(context).textTheme.titleLarge),
-                  const SizedBox(height: 10),
-                  Text('• ${fasting.start} hours fasting',
-                      style: Theme.of(context).textTheme.bodyMedium),
-                  Text('• ${fasting.end} hours eating period',
-                      style: Theme.of(context).textTheme.bodyMedium),
+                  Text(
+                    fasting.title,
+                    style: Theme.of(context).textTheme.titleLarge?.copyWith(
+                      fontWeight: FontWeight.w700,
+                      color: Colors.white,
+                    ),
+                  ),
+                  const SizedBox(height: 12),
+                  Text(
+                    '${fasting.start} hours fasting',
+                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                      color: Colors.white.withOpacity(0.9),
+                      fontWeight: FontWeight.w500,
+                    ),
+                  ),
+                  const SizedBox(height: 4),
+                  Text(
+                    '${fasting.end} hours eating period',
+                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                      color: Colors.white.withOpacity(0.9),
+                      fontWeight: FontWeight.w500,
+                    ),
+                  ),
                 ],
               ),
             ),
-            Icon(icon, size: 50)
+            Container(
+              width: 60,
+              height: 60,
+              decoration: BoxDecoration(
+                color: Colors.white.withOpacity(0.2),
+                shape: BoxShape.circle,
+              ),
+              child: Icon(
+                icon,
+                size: 30,
+                color: Colors.white,
+              ),
+            ),
           ],
         ),
       ),
