@@ -3,39 +3,78 @@ import 'package:dartz/dartz.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logger/logger.dart';
+import 'package:google_sign_in/google_sign_in.dart';
+import 'package:sign_in_with_apple/sign_in_with_apple.dart';
+import 'package:crypto/crypto.dart';
+import 'dart:convert';
+import 'dart:math';
 
 class AuthService {
   final FirebaseAuth _firebaseAuth;
 
   AuthService(this._firebaseAuth);
 
-  Future<Either<String, User>> signup(
-      {required String email, required String password}) async {
+  // Removed email/password authentication methods
+  // Only Google and Apple Sign-In are supported now
+
+  Future<Either<String, User?>> signInWithGoogle() async {
     try {
-      final response = await _firebaseAuth.createUserWithEmailAndPassword(
-          email: email, password: password);
-      return right(response.user!);
-    } on FirebaseAuthException catch (e) {
-      return left(e.message ?? 'Failed to Signup.');
+      final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
+      if (googleUser == null) {
+        return left('Google Sign-In was cancelled');
+      }
+
+      final GoogleSignInAuthentication googleAuth = await googleUser.authentication;
+      final credential = GoogleAuthProvider.credential(
+        accessToken: googleAuth.accessToken,
+        idToken: googleAuth.idToken,
+      );
+
+      final UserCredential result = await _firebaseAuth.signInWithCredential(credential);
+      return right(result.user);
+    } catch (e) {
+      Logger().e('Google Sign-In error: $e');
+      return left('Failed to sign in with Google: ${e.toString()}');
     }
   }
 
-  Future<Either<String, User?>> login(
-      {required String email, required String password}) async {
+  Future<Either<String, User?>> signInWithApple() async {
     try {
-      final response = await _firebaseAuth.signInWithEmailAndPassword(
-        email: email,
-        password: password,
+      final rawNonce = _generateNonce();
+      final nonce = _sha256ofString(rawNonce);
+
+      final appleCredential = await SignInWithApple.getAppleIDCredential(
+        scopes: [AppleIDAuthorizationScopes.email, AppleIDAuthorizationScopes.fullName],
+        nonce: nonce,
       );
-      Logger().d('result $response');
-      return right(response.user);
-    } on FirebaseAuthException catch (e) {
-      Logger().e('result e $e');
-      return left(e.message ?? 'Failed to Login');
+
+      final oauthCredential = OAuthProvider('apple.com').credential(
+        idToken: appleCredential.identityToken,
+        rawNonce: rawNonce,
+      );
+
+      final UserCredential result = await _firebaseAuth.signInWithCredential(oauthCredential);
+      return right(result.user);
+    } catch (e) {
+      Logger().e('Apple Sign-In error: $e');
+      return left('Failed to sign in with Apple: ${e.toString()}');
     }
+  }
+
+  String _generateNonce([int length = 32]) {
+    const charset = '0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._';
+    final random = Random.secure();
+    return List.generate(length, (_) => charset[random.nextInt(charset.length)]).join();
+  }
+
+  String _sha256ofString(String input) {
+    final bytes = utf8.encode(input);
+    final digest = sha256.convert(bytes);
+    return digest.toString();
   }
 
   Future<void> signout() async {
+    await GoogleSignIn().signOut();
     return await _firebaseAuth.signOut();
   }
 }
