@@ -19,6 +19,8 @@ final firebaseAuthProvider = Provider<FirebaseAuth>.internal(
   allTransitiveDependencies: null,
 );
 
+@Deprecated('Will be removed in 3.0. Use Ref instead')
+// ignore: unused_element
 typedef FirebaseAuthRef = ProviderRef<FirebaseAuth>;
 String _$firebaseFirestoreHash() => r'a5daf1e4ce5c8fb23ae8553b86881c5518f41d4d';
 
@@ -34,6 +36,8 @@ final firebaseFirestoreProvider = Provider<FirebaseFirestore>.internal(
   allTransitiveDependencies: null,
 );
 
+@Deprecated('Will be removed in 3.0. Use Ref instead')
+// ignore: unused_element
 typedef FirebaseFirestoreRef = ProviderRef<FirebaseFirestore>;
 String _$firebaseStorageHash() => r'9b239ccbfd0ef129cc8465924f751c6af8cb3563';
 
@@ -49,6 +53,8 @@ final firebaseStorageProvider = Provider<FirebaseStorage>.internal(
   allTransitiveDependencies: null,
 );
 
+@Deprecated('Will be removed in 3.0. Use Ref instead')
+// ignore: unused_element
 typedef FirebaseStorageRef = ProviderRef<FirebaseStorage>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
