@@ -2,7 +2,6 @@ import 'package:awaku/service/model/fasting_model.dart';
 import 'package:awaku/service/provider/fasting_provider.dart';
 import 'package:awaku/service/stop_watch_service.dart';
 import 'package:awaku/utils/extensions.dart';
-import 'package:awaku/widgets/custom_button.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -12,9 +11,9 @@ import 'package:stop_watch_timer/stop_watch_timer.dart';
 class FastingDetail extends ConsumerStatefulWidget {
   final FastingModel fasting;
   const FastingDetail({
-    Key? key,
+    super.key,
     required this.fasting,
-  }) : super(key: key);
+  });
 
   @override
   ConsumerState<FastingDetail> createState() => _FastingDetailState();
@@ -38,88 +37,322 @@ class _FastingDetailState extends ConsumerState<FastingDetail> {
 
   @override
   Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+    
     return Scaffold(
-      appBar: AppBar(),
+      backgroundColor: colorScheme.surface,
+      appBar: AppBar(
+        title: Text(
+          widget.fasting.title,
+          style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+            fontWeight: FontWeight.w700,
+            color: colorScheme.onSurface,
+          ),
+        ),
+        backgroundColor: colorScheme.surface,
+        elevation: 0,
+        scrolledUnderElevation: 0,
+        leading: IconButton(
+          onPressed: () => Navigator.of(context).pop(),
+          icon: Icon(
+            Icons.arrow_back_ios_rounded,
+            color: colorScheme.onSurface,
+          ),
+        ),
+      ),
       body: SingleChildScrollView(
+        padding: const EdgeInsets.all(24),
         child: Column(
           children: [
-            Text(widget.fasting.title,
-                style: Theme.of(context).textTheme.displaySmall),
-            const Text('Fasting'),
-            if (stopWatchTimer.isRunning)
-              Column(
+            // Header Section
+            Container(
+              width: double.infinity,
+              padding: const EdgeInsets.all(32),
+              decoration: BoxDecoration(
+                gradient: LinearGradient(
+                  colors: [
+                    const Color(0xFF667EEA).withOpacity(0.1),
+                    const Color(0xFF764BA2).withOpacity(0.05),
+                  ],
+                  begin: Alignment.topLeft,
+                  end: Alignment.bottomRight,
+                ),
+                borderRadius: BorderRadius.circular(24),
+                border: Border.all(
+                  color: const Color(0xFF667EEA).withOpacity(0.2),
+                  width: 1,
+                ),
+              ),
+              child: Column(
                 children: [
-                  StreamBuilder<int>(
-                    stream: stopWatchTimer.rawTime,
-                    initialData: stopWatchTimer.rawTime.value,
-                    builder: (context, snap) {
-                      final value = snap.data!;
-                      final displayTime =
-                          StopWatchTimer.getDisplayTime(value, hours: true);
-                      return Column(
-                        children: <Widget>[
-                          Padding(
-                            padding: const EdgeInsets.all(8),
-                            child: Text(
-                              displayTime,
-                              style: Theme.of(context).textTheme.displayMedium,
+                  Container(
+                    width: 100,
+                    height: 100,
+                    decoration: BoxDecoration(
+                      gradient: const LinearGradient(
+                        colors: [Color(0xFF667EEA), Color(0xFF764BA2)],
+                        begin: Alignment.topLeft,
+                        end: Alignment.bottomRight,
+                      ),
+                      shape: BoxShape.circle,
+                      boxShadow: [
+                        BoxShadow(
+                          color: const Color(0xFF667EEA).withOpacity(0.3),
+                          blurRadius: 20,
+                          offset: const Offset(0, 8),
+                        ),
+                      ],
+                    ),
+                    child: const Icon(
+                      Icons.schedule_rounded,
+                      size: 50,
+                      color: Colors.white,
+                    ),
+                  ),
+                  const SizedBox(height: 24),
+                  Text(
+                    'Intermittent Fasting',
+                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                      color: colorScheme.onSurface.withOpacity(0.7),
+                      fontWeight: FontWeight.w500,
+                    ),
+                  ),
+                  const SizedBox(height: 8),
+                  Text(
+                    '${widget.fasting.start}:${widget.fasting.end} Schedule',
+                    style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+                      fontWeight: FontWeight.w700,
+                      color: colorScheme.onSurface,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            const SizedBox(height: 32),
+            
+            // Timer Section
+            if (stopWatchTimer.isRunning)
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(32),
+                decoration: BoxDecoration(
+                  color: colorScheme.surfaceContainer,
+                  borderRadius: BorderRadius.circular(24),
+                  border: Border.all(
+                    color: colorScheme.outline.withOpacity(0.1),
+                    width: 1,
+                  ),
+                ),
+                child: Column(
+                  children: [
+                    Text(
+                      'Fasting in Progress',
+                      style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                        color: const Color(0xFF10B981),
+                        fontWeight: FontWeight.w600,
+                      ),
+                    ),
+                    const SizedBox(height: 24),
+                    StreamBuilder<int>(
+                      stream: stopWatchTimer.rawTime,
+                      initialData: stopWatchTimer.rawTime.value,
+                      builder: (context, snap) {
+                        final value = snap.data!;
+                        final displayTime =
+                            StopWatchTimer.getDisplayTime(value, hours: true);
+                        return Container(
+                          padding: const EdgeInsets.symmetric(
+                            horizontal: 32,
+                            vertical: 24,
+                          ),
+                          decoration: BoxDecoration(
+                            gradient: LinearGradient(
+                              colors: [
+                                const Color(0xFF10B981).withOpacity(0.1),
+                                const Color(0xFF059669).withOpacity(0.05),
+                              ],
+                              begin: Alignment.topLeft,
+                              end: Alignment.bottomRight,
+                            ),
+                            borderRadius: BorderRadius.circular(20),
+                            border: Border.all(
+                              color: const Color(0xFF10B981).withOpacity(0.2),
+                              width: 1,
                             ),
                           ),
+                          child: Text(
+                            displayTime,
+                            style: Theme.of(context).textTheme.displayLarge?.copyWith(
+                              fontWeight: FontWeight.w700,
+                              color: const Color(0xFF10B981),
+                              fontFeatures: [const FontFeature.tabularFigures()],
+                            ),
+                          ),
+                        );
+                      },
+                    ),
+                    const SizedBox(height: 24),
+                    Container(
+                      padding: const EdgeInsets.all(20),
+                      decoration: BoxDecoration(
+                        color: colorScheme.surface,
+                        borderRadius: BorderRadius.circular(16),
+                        border: Border.all(
+                          color: colorScheme.outline.withOpacity(0.1),
+                          width: 1,
+                        ),
+                      ),
+                      child: Row(
+                        children: [
+                          Icon(
+                            Icons.flag_rounded,
+                            color: colorScheme.primary,
+                            size: 20,
+                          ),
+                          const SizedBox(width: 12),
+                          Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                'Expected End Time',
+                                style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                                  color: colorScheme.onSurface.withOpacity(0.7),
+                                  fontWeight: FontWeight.w500,
+                                ),
+                              ),
+                              const SizedBox(height: 4),
+                              Text(
+                                formatTime.format(end),
+                                style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                                  fontWeight: FontWeight.w600,
+                                  color: colorScheme.onSurface,
+                                ),
+                              ),
+                            ],
+                          ),
                         ],
-                      );
-                    },
-                  ),
-                  const Text('End (expected)'),
-                  Text(formatTime.format(end)),
-                ],
+                      ),
+                    ),
+                  ],
+                ),
               )
-            else ...[
-              ListTile(
-                minLeadingWidth: 10,
-                leading: const Icon(
-                  Icons.circle,
-                  size: 15,
-                  color: Colors.green,
+            else
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(24),
+                decoration: BoxDecoration(
+                  color: colorScheme.surfaceContainer,
+                  borderRadius: BorderRadius.circular(20),
+                  border: Border.all(
+                    color: colorScheme.outline.withOpacity(0.1),
+                    width: 1,
+                  ),
                 ),
-                title: const Text('Start'),
-                subtitle: Text(formatDayTime.format(start)),
-                trailing: const Icon(Icons.edit),
-                onTap: showTimePicker,
-              ),
-              ListTile(
-                minLeadingWidth: 10,
-                leading: const Icon(
-                  Icons.circle,
-                  size: 15,
-                  color: Colors.red,
+                child: Column(
+                  children: [
+                    Text(
+                      'Schedule Your Fast',
+                      style: Theme.of(context).textTheme.titleLarge?.copyWith(
+                        fontWeight: FontWeight.w600,
+                        color: colorScheme.onSurface,
+                      ),
+                    ),
+                    const SizedBox(height: 24),
+                    _buildTimeCard(
+                      context,
+                      icon: Icons.play_circle_outline_rounded,
+                      iconColor: const Color(0xFF10B981),
+                      title: 'Start Time',
+                      time: formatDayTime.format(start),
+                      onTap: showTimePicker,
+                      showEdit: true,
+                    ),
+                    const SizedBox(height: 16),
+                    _buildTimeCard(
+                      context,
+                      icon: Icons.flag_rounded,
+                      iconColor: const Color(0xFFEF4444),
+                      title: 'Expected End',
+                      time: formatDayTime.format(end),
+                      onTap: null,
+                      showEdit: false,
+                    ),
+                  ],
                 ),
-                title: const Text('End (expected)'),
-                subtitle: Text(formatDayTime.format(end)),
               ),
-            ],
-            Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: CustomButton(
-                  backgroundColor: Colors.blue,
-                  width: double.infinity,
-                  isDisabled: false,
-                  title: stopWatchTimer.isRunning
-                      ? 'End fasting'
-                      : 'Start fasting',
-                  onPressed: () {
+            const SizedBox(height: 32),
+            // Action Button
+            Container(
+              width: double.infinity,
+              height: 56,
+              decoration: BoxDecoration(
+                gradient: stopWatchTimer.isRunning
+                    ? const LinearGradient(
+                        colors: [Color(0xFFEF4444), Color(0xFFDC2626)],
+                        begin: Alignment.topLeft,
+                        end: Alignment.bottomRight,
+                      )
+                    : const LinearGradient(
+                        colors: [Color(0xFF10B981), Color(0xFF059669)],
+                        begin: Alignment.topLeft,
+                        end: Alignment.bottomRight,
+                      ),
+                borderRadius: BorderRadius.circular(16),
+                boxShadow: [
+                  BoxShadow(
+                    color: (stopWatchTimer.isRunning
+                            ? const Color(0xFFEF4444)
+                            : const Color(0xFF10B981))
+                        .withOpacity(0.3),
+                    blurRadius: 12,
+                    offset: const Offset(0, 6),
+                  ),
+                ],
+              ),
+              child: Material(
+                color: Colors.transparent,
+                child: InkWell(
+                  borderRadius: BorderRadius.circular(16),
+                  onTap: () {
                     if (stopWatchTimer.isRunning) {
                       _showEndDialog(context);
                     } else {
                       stopWatchTimer.clearPresetTime();
                       stopWatchTimer.setPresetHoursTime(widget.fasting.start);
                       stopWatchTimer.onStartTimer();
-                      ref.read(startFastingProvider.notifier).set(true);
+                      ref.read(startFastingProvider.notifier).state = true;
                       ref
                           .read(selectedFastingProvider.notifier)
-                          .set(widget.fasting);
+                          .state = widget.fasting;
                     }
                     setState(() {});
-                  }),
+                  },
+                  child: Center(
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        Icon(
+                          stopWatchTimer.isRunning
+                              ? Icons.stop_circle_rounded
+                              : Icons.play_circle_rounded,
+                          color: Colors.white,
+                          size: 24,
+                        ),
+                        const SizedBox(width: 12),
+                        Text(
+                          stopWatchTimer.isRunning
+                              ? 'End Fasting'
+                              : 'Start Fasting',
+                          style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                            color: Colors.white,
+                            fontWeight: FontWeight.w600,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
             ),
           ],
         ),
@@ -170,6 +403,81 @@ class _FastingDetailState extends ConsumerState<FastingDetail> {
     );
   }
 
+  Widget _buildTimeCard(
+    BuildContext context, {
+    required IconData icon,
+    required Color iconColor,
+    required String title,
+    required String time,
+    required VoidCallback? onTap,
+    required bool showEdit,
+  }) {
+    final colorScheme = Theme.of(context).colorScheme;
+    
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: colorScheme.surface,
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: colorScheme.outline.withOpacity(0.1),
+          width: 1,
+        ),
+      ),
+      child: InkWell(
+        borderRadius: BorderRadius.circular(16),
+        onTap: onTap,
+        child: Row(
+          children: [
+            Container(
+              width: 48,
+              height: 48,
+              decoration: BoxDecoration(
+                color: iconColor.withOpacity(0.1),
+                shape: BoxShape.circle,
+              ),
+              child: Icon(
+                icon,
+                color: iconColor,
+                size: 24,
+              ),
+            ),
+            const SizedBox(width: 16),
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    title,
+                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                      color: colorScheme.onSurface.withOpacity(0.7),
+                      fontWeight: FontWeight.w500,
+                    ),
+                  ),
+                  const SizedBox(height: 4),
+                  Text(
+                    time,
+                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                      fontWeight: FontWeight.w600,
+                      color: colorScheme.onSurface,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            if (showEdit)
+              Icon(
+                Icons.edit_rounded,
+                color: colorScheme.onSurface.withOpacity(0.5),
+                size: 20,
+              ),
+          ],
+        ),
+      ),
+    );
+  }
+
   void _showEndDialog(BuildContext context) {
     showCupertinoModalPopup<void>(
       context: context,
@@ -189,7 +497,7 @@ class _FastingDetailState extends ConsumerState<FastingDetail> {
             onPressed: () async {
               stopWatchTimer.onStopTimer();
               stopWatchTimer.onResetTimer();
-              ref.read(startFastingProvider.notifier).set(false);
+              ref.read(startFastingProvider.notifier).state = false;
               if (mounted) Navigator.pop(context);
             },
             child: const Text('Yes'),
