@@ -1,14 +1,12 @@
 import 'package:awaku/service/provider/health_provider.dart';
 import 'package:awaku/service/provider/profile_provider.dart';
-import 'package:awaku/service/provider/states/profile_states.dart';
+import 'package:awaku/src/localization/app_localizations.dart';
 import 'package:awaku/utils/extensions.dart';
 import 'package:awaku/utils/validator.dart';
 import 'package:awaku/widgets/custom_button.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
@@ -20,7 +18,6 @@ class ProfileView extends ConsumerStatefulWidget {
 }
 
 class _ProfileViewState extends ConsumerState<ProfileView> {
-  User? user = FirebaseAuth.instance.currentUser;
   final GlobalKey<FormState> _formPersonal = GlobalKey();
   TextEditingController name = TextEditingController();
   TextEditingController height = TextEditingController();
@@ -47,184 +44,285 @@ class _ProfileViewState extends ConsumerState<ProfileView> {
 
   @override
   Widget build(BuildContext context) {
-    ref.listen(profileProvider, (previous, next) {
-      if (next is ProfileStateError) {
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          content: Text(next.error),
-          behavior: SnackBarBehavior.floating,
-        ));
-      } else if (next is ProfileStateSuccess) {
-        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-          content: Text('Changes updated'),
-          behavior: SnackBarBehavior.floating,
-        ));
-      }
-    });
+    final colorScheme = Theme.of(context).colorScheme;
+    
     return Scaffold(
+      backgroundColor: colorScheme.surface,
       appBar: AppBar(
-        title: Text(AppLocalizations.of(context)!.personalData),
+        title: Text(
+          AppLocalizations.of(context)!.personalData,
+          style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+            fontWeight: FontWeight.w700,
+            color: colorScheme.onSurface,
+          ),
+        ),
+        backgroundColor: colorScheme.surface,
+        elevation: 0,
+        scrolledUnderElevation: 0,
+        leading: IconButton(
+          onPressed: () => Navigator.of(context).pop(),
+          icon: Icon(
+            Icons.arrow_back_ios_rounded,
+            color: colorScheme.onSurface,
+          ),
+        ),
       ),
       body: Form(
         key: _formPersonal,
         child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.symmetric(vertical: 8.0),
-                  child: Text(
-                    AppLocalizations.of(context)!.yourName,
-                    style: Theme.of(context).textTheme.titleMedium,
-                  ),
-                ),
-                TextFormField(
-                  controller: name,
-                  decoration: InputDecoration(
-                    fillColor: Colors.blueGrey[50],
-                    filled: true,
-                    border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(5.0),
-                        borderSide: BorderSide.none),
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.symmetric(vertical: 8.0),
-                  child: Text(
-                    AppLocalizations.of(context)!.gender,
-                    style: Theme.of(context).textTheme.titleMedium,
-                  ),
-                ),
-                DropdownButtonFormField(
-                    isExpanded: true,
-                    decoration: InputDecoration(
-                      fillColor: Colors.blueGrey[50],
-                      filled: true,
-                      border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(5.0),
-                          borderSide: BorderSide.none),
-                    ),
-                    items: const [
-                      DropdownMenuItem(value: 'male', child: Text("Male")),
-                      DropdownMenuItem(value: 'female', child: Text("Female")),
+          padding: const EdgeInsets.all(24),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              // Profile Header
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(24),
+                decoration: BoxDecoration(
+                  gradient: LinearGradient(
+                    colors: [
+                      colorScheme.primary.withOpacity(0.1),
+                      colorScheme.primary.withOpacity(0.05),
                     ],
-                    hint: const Text("Select City"),
-                    value: gender,
-                    onChanged: (val) {
-                      setState(() {
-                        gender = val!;
-                      });
-                    }),
-                const SizedBox(height: 10),
-                Padding(
-                  padding: const EdgeInsets.symmetric(vertical: 8.0),
-                  child: Text(
-                    AppLocalizations.of(context)!.dateOfBirth,
-                    style: Theme.of(context).textTheme.titleMedium,
+                    begin: Alignment.topLeft,
+                    end: Alignment.bottomRight,
+                  ),
+                  borderRadius: BorderRadius.circular(20),
+                  border: Border.all(
+                    color: colorScheme.primary.withOpacity(0.1),
+                    width: 1,
                   ),
                 ),
-                Container(
-                  height: 60,
-                  decoration: BoxDecoration(
-                    color: Colors.blueGrey[50],
-                    borderRadius: BorderRadius.circular(5.0),
-                  ),
-                  child: Center(
-                    child: ListTile(
-                      title: Text(dob != null ? formateDate.format(dob!) : ''),
-                      onTap: () => showBottom(),
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 10),
-                Row(
+                child: Column(
                   children: [
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Padding(
-                            padding: const EdgeInsets.symmetric(vertical: 8.0),
-                            child: Text(
-                              AppLocalizations.of(context)!.height,
-                              style: Theme.of(context).textTheme.titleMedium,
-                            ),
-                          ),
-                          TextFormField(
-                            controller: height,
-                            validator: heightRequired,
-                            keyboardType: TextInputType.number,
-                            inputFormatters: <TextInputFormatter>[
-                              FilteringTextInputFormatter.digitsOnly
-                            ],
-                            decoration: InputDecoration(
-                              fillColor: Colors.blueGrey[50],
-                              filled: true,
-                              border: OutlineInputBorder(
-                                  borderRadius: BorderRadius.circular(5.0),
-                                  borderSide: BorderSide.none),
-                            ),
-                          ),
-                        ],
+                    Container(
+                      width: 80,
+                      height: 80,
+                      decoration: BoxDecoration(
+                        color: colorScheme.primary.withOpacity(0.1),
+                        shape: BoxShape.circle,
+                        border: Border.all(
+                          color: colorScheme.primary.withOpacity(0.2),
+                          width: 2,
+                        ),
+                      ),
+                      child: Icon(
+                        Icons.person_rounded,
+                        size: 40,
+                        color: colorScheme.primary,
                       ),
                     ),
-                    const SizedBox(width: 10),
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Padding(
-                            padding: const EdgeInsets.symmetric(vertical: 8.0),
-                            child: Text(
-                              AppLocalizations.of(context)!.weight,
-                              style: Theme.of(context).textTheme.titleMedium,
-                            ),
-                          ),
-                          TextFormField(
-                            controller: weight,
-                            validator: weightRequired,
-                            keyboardType: TextInputType.number,
-                            inputFormatters: <TextInputFormatter>[
-                              FilteringTextInputFormatter.allow(
-                                  RegExp(r'^\d*\.?\d{0,2}')),
-                            ],
-                            decoration: InputDecoration(
-                              fillColor: Colors.blueGrey[50],
-                              filled: true,
-                              border: OutlineInputBorder(
-                                  borderRadius: BorderRadius.circular(5.0),
-                                  borderSide: BorderSide.none),
-                            ),
-                          )
-                        ],
+                    const SizedBox(height: 16),
+                    Text(
+                      'Personal Information',
+                      style: Theme.of(context).textTheme.titleLarge?.copyWith(
+                        fontWeight: FontWeight.w600,
+                        color: colorScheme.onSurface,
                       ),
-                    )
+                    ),
+                    const SizedBox(height: 8),
+                    Text(
+                      'Update your profile details',
+                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                        color: colorScheme.onSurface.withOpacity(0.7),
+                      ),
+                    ),
                   ],
                 ),
-                const SizedBox(height: 20),
-                CustomButton(
-                  width: double.infinity,
-                  isDisabled: false,
-                  title: AppLocalizations.of(context)!.saveChanges,
-                  onPressed: () async {
-                    if (_formPersonal.currentState!.validate()) {
-                      ref.read(profileProvider.notifier).update(
-                            uid: user!.uid,
+              ),
+              const SizedBox(height: 32),
+              
+              // Form Fields
+              _buildInputField(
+                context,
+                colorScheme,
+                label: AppLocalizations.of(context)!.yourName,
+                icon: Icons.person_outline_rounded,
+                child: TextFormField(
+                  controller: name,
+                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                    color: colorScheme.onSurface,
+                  ),
+                  decoration: _buildInputDecoration(context, colorScheme, 'Enter your full name'),
+                ),
+              ),
+              const SizedBox(height: 24),
+              
+              _buildInputField(
+                context,
+                colorScheme,
+                label: AppLocalizations.of(context)!.gender,
+                icon: Icons.wc_rounded,
+                child: DropdownButtonFormField<String>(
+                  value: gender,
+                  isExpanded: true,
+                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                    color: colorScheme.onSurface,
+                  ),
+                  decoration: _buildInputDecoration(context, colorScheme, 'Select gender'),
+                  items: const [
+                    DropdownMenuItem(value: 'male', child: Text('Male')),
+                    DropdownMenuItem(value: 'female', child: Text('Female')),
+                  ],
+                  onChanged: (val) {
+                    setState(() {
+                      gender = val!;
+                    });
+                  },
+                ),
+              ),
+              const SizedBox(height: 24),
+              
+              _buildInputField(
+                context,
+                colorScheme,
+                label: AppLocalizations.of(context)!.dateOfBirth,
+                icon: Icons.calendar_today_rounded,
+                child: Container(
+                  decoration: BoxDecoration(
+                    color: colorScheme.surfaceContainerLow,
+                    borderRadius: BorderRadius.circular(16),
+                    border: Border.all(
+                      color: colorScheme.outline.withOpacity(0.2),
+                      width: 1,
+                    ),
+                  ),
+                  child: Material(
+                    color: Colors.transparent,
+                    child: InkWell(
+                      onTap: () => showBottom(),
+                      borderRadius: BorderRadius.circular(16),
+                      child: Padding(
+                        padding: const EdgeInsets.all(16),
+                        child: Row(
+                          children: [
+                            Expanded(
+                              child: Text(
+                                dob != null ? formateDate.format(dob!) : 'Select date of birth',
+                                style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                                  color: dob != null 
+                                      ? colorScheme.onSurface 
+                                      : colorScheme.onSurface.withOpacity(0.5),
+                                ),
+                              ),
+                            ),
+                            Icon(
+                              Icons.arrow_drop_down_rounded,
+                              color: colorScheme.onSurface.withOpacity(0.5),
+                            ),
+                          ],
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 24),
+              
+              // Height and Weight Row
+              Row(
+                children: [
+                  Expanded(
+                    child: _buildInputField(
+                      context,
+                      colorScheme,
+                      label: AppLocalizations.of(context)!.height,
+                      icon: Icons.height_rounded,
+                      child: TextFormField(
+                        controller: height,
+                        validator: heightRequired,
+                        keyboardType: TextInputType.number,
+                        inputFormatters: <TextInputFormatter>[
+                          FilteringTextInputFormatter.digitsOnly
+                        ],
+                        style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                          color: colorScheme.onSurface,
+                        ),
+                        decoration: _buildInputDecoration(context, colorScheme, 'cm'),
+                      ),
+                    ),
+                  ),
+                  const SizedBox(width: 16),
+                  Expanded(
+                    child: _buildInputField(
+                      context,
+                      colorScheme,
+                      label: AppLocalizations.of(context)!.weight,
+                      icon: Icons.monitor_weight_outlined,
+                      child: TextFormField(
+                        controller: weight,
+                        validator: weightRequired,
+                        keyboardType: TextInputType.number,
+                        inputFormatters: <TextInputFormatter>[
+                          FilteringTextInputFormatter.allow(
+                              RegExp(r'^\d*\.?\d{0,2}')),
+                        ],
+                        style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                          color: colorScheme.onSurface,
+                        ),
+                        decoration: _buildInputDecoration(context, colorScheme, 'kg'),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+              const SizedBox(height: 40),
+              
+              // Save Button
+              Container(
+                width: double.infinity,
+                height: 56,
+                decoration: BoxDecoration(
+                  gradient: LinearGradient(
+                    colors: [colorScheme.primary, colorScheme.primary.withOpacity(0.8)],
+                    begin: Alignment.topLeft,
+                    end: Alignment.bottomRight,
+                  ),
+                  borderRadius: BorderRadius.circular(16),
+                  boxShadow: [
+                    BoxShadow(
+                      color: colorScheme.primary.withOpacity(0.3),
+                      blurRadius: 12,
+                      offset: const Offset(0, 4),
+                    ),
+                  ],
+                ),
+                child: Material(
+                  color: Colors.transparent,
+                  child: InkWell(
+                    onTap: () async {
+                      if (_formPersonal.currentState!.validate()) {
+                        final profile = ref.read(fetchUserProvider);
+                        if (profile.hasValue && profile.value?.uid != null) {
+                          await updateProfile(ref,
+                            uid: profile.value!.uid!,
                             name: name.text,
                             dob: dob,
                             weight: double.parse(weight.text),
                             height: int.parse(height.text),
                             gender: gender,
                           );
-                      await ref.read(healthProvider).addWeightAndHeight(
-                          weight: double.parse(weight.text),
-                          height: double.parse(height.text));
-                    }
-                  },
+                          await ref.read(healthProvider).addWeightAndHeight(
+                              weight: double.parse(weight.text),
+                              height: double.parse(height.text));
+                        }
+                      }
+                    },
+                    borderRadius: BorderRadius.circular(16),
+                    child: Center(
+                      child: Text(
+                        AppLocalizations.of(context)!.saveChanges,
+                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                          color: colorScheme.onPrimary,
+                          fontWeight: FontWeight.w600,
+                        ),
+                      ),
+                    ),
+                  ),
                 ),
-              ],
-            ),
+              ),
+              const SizedBox(height: 24),
+            ],
           ),
         ),
       ),
@@ -267,6 +365,97 @@ class _ProfileViewState extends ConsumerState<ProfileView> {
           ),
         );
       },
+    );
+  }
+
+  Widget _buildInputField(
+    BuildContext context,
+    ColorScheme colorScheme, {
+    required String label,
+    required IconData icon,
+    required Widget child,
+  }) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Row(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(8),
+              decoration: BoxDecoration(
+                color: colorScheme.primary.withOpacity(0.1),
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: Icon(
+                icon,
+                size: 20,
+                color: colorScheme.primary,
+              ),
+            ),
+            const SizedBox(width: 12),
+            Text(
+              label,
+              style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                fontWeight: FontWeight.w600,
+                color: colorScheme.onSurface,
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(height: 12),
+        child,
+      ],
+    );
+  }
+
+  InputDecoration _buildInputDecoration(
+    BuildContext context,
+    ColorScheme colorScheme,
+    String hintText,
+  ) {
+    return InputDecoration(
+      hintText: hintText,
+      hintStyle: Theme.of(context).textTheme.bodyLarge?.copyWith(
+        color: colorScheme.onSurface.withOpacity(0.5),
+      ),
+      filled: true,
+      fillColor: colorScheme.surfaceContainerLow,
+      border: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(
+          color: colorScheme.outline.withOpacity(0.2),
+          width: 1,
+        ),
+      ),
+      enabledBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(
+          color: colorScheme.outline.withOpacity(0.2),
+          width: 1,
+        ),
+      ),
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(
+          color: colorScheme.primary,
+          width: 2,
+        ),
+      ),
+      errorBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(
+          color: colorScheme.error,
+          width: 1,
+        ),
+      ),
+      focusedErrorBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(
+          color: colorScheme.error,
+          width: 2,
+        ),
+      ),
+      contentPadding: const EdgeInsets.all(16),
     );
   }
 }
