@@ -1,9 +1,8 @@
 import 'package:awaku/service/provider/authentication_provider.dart';
 import 'package:awaku/service/provider/health_provider.dart';
 import 'package:awaku/service/provider/profile_provider.dart';
-import 'package:awaku/service/provider/states/profile_states.dart';
+import 'package:awaku/src/localization/app_localizations.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
@@ -36,114 +35,69 @@ class _SettingsViewState extends ConsumerState<SettingsView> {
   @override
   Widget build(BuildContext context) {
     final user = ref.watch(fetchUserProvider);
-    final healthAuth = ref.watch(healthAuthNotifierProvider);
-    ref.listen(profileProvider, (previous, next) {
-      if (next is ProfileStateError) {
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          content: Text(next.error),
-          behavior: SnackBarBehavior.floating,
-        ));
-      } else if (next is ProfileStateSuccess) {
-        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-          content: Text('Changes updated'),
-          behavior: SnackBarBehavior.floating,
-        ));
-        ref.invalidate(fetchUserProvider);
-      }
-    });
+    final healthAuth = ref.watch(healthAuthProvider);
+    final colorScheme = Theme.of(context).colorScheme;
+    
     return Scaffold(
-        appBar: AppBar(
-          title: Text(AppLocalizations.of(context)!.settings),
+      backgroundColor: colorScheme.surface,
+      appBar: AppBar(
+        title: Text(
+          AppLocalizations.of(context)!.settings,
+          style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+            fontWeight: FontWeight.w700,
+            color: colorScheme.onSurface,
+          ),
         ),
-        body: Column(
+        backgroundColor: colorScheme.surface,
+        elevation: 0,
+        scrolledUnderElevation: 0,
+        leading: IconButton(
+          onPressed: () => Navigator.of(context).pop(),
+          icon: Icon(
+            Icons.arrow_back_ios_rounded,
+            color: colorScheme.onSurface,
+          ),
+        ),
+      ),
+      body: SingleChildScrollView(
+        padding: const EdgeInsets.all(24),
+        child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            ListTile(
-              leading: const Icon(Icons.account_box_outlined),
-              title: Text(AppLocalizations.of(context)!.personalData),
-              trailing: const Icon(Icons.arrow_forward_ios_outlined),
-              onTap: () => context.push('/setting/profile'),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 16.0, left: 16.0),
-              child: Text(
-                AppLocalizations.of(context)!.feature,
-                style: Theme.of(context).textTheme.titleMedium,
-              ),
-            ),
-            ListTile(
-              leading: const Icon(Icons.local_dining),
-              title: Text(AppLocalizations.of(context)!.intermittentFasting),
-              trailing: Switch(
-                  value: enableFasting,
-                  onChanged: (val) {
-                    setState(() => enableFasting = !enableFasting);
-                    ref.read(profileProvider.notifier).update(
-                          uid: user.value!.uid!,
-                          enableFasting: val,
-                        );
-                  }),
-            ),
-            ListTile(
-              leading: const Icon(Icons.local_drink),
-              title: Text(AppLocalizations.of(context)!.water),
-              trailing: Switch(
-                  value: enableWater,
-                  onChanged: (val) {
-                    setState(() => enableWater = !enableWater);
-                    ref.read(profileProvider.notifier).update(
-                          uid: user.value!.uid!,
-                          enableWater: val,
-                        );
-                  }),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 16.0, left: 16.0),
-              child: Text(
-                AppLocalizations.of(context)!.tools,
-                style: Theme.of(context).textTheme.titleMedium,
-              ),
-            ),
-            ListTile(
-              leading: const Icon(Icons.device_hub),
-              title: Text(AppLocalizations.of(context)!.devices),
-              trailing: const Icon(Icons.arrow_forward_ios_outlined),
-              onTap: () => context.push('/setting/add-device'),
-            ),
-            ListTile(
-              leading: const Icon(Icons.healing_outlined),
-              title: Text(AppLocalizations.of(context)!.health),
-              trailing: healthAuth.value == true
-                  ? Text(
-                      AppLocalizations.of(context)!.connected,
-                      style: Theme.of(context)
-                          .textTheme
-                          .bodyMedium!
-                          .copyWith(color: Colors.green),
-                    )
-                  : ElevatedButton(
-                      onPressed: requestAuthHealt,
-                      child: Text(
-                        AppLocalizations.of(context)!.requestAuth,
-                      )),
-            ),
-            ListTile(
-              title: Text(AppLocalizations.of(context)!.logout),
-              trailing: const Icon(Icons.logout),
-              onTap: () {
-                ref.read(authenticationProvider.notifier).logout();
-              },
-            ),
+            // Profile Section
+            _buildProfileSection(context, colorScheme),
+            const SizedBox(height: 32),
+            
+            // Features Section
+            _buildSectionHeader(context, AppLocalizations.of(context)!.feature, colorScheme),
+            const SizedBox(height: 16),
+            _buildFeaturesSection(context, user, colorScheme),
+            const SizedBox(height: 32),
+            
+            // Tools Section
+            _buildSectionHeader(context, AppLocalizations.of(context)!.tools, colorScheme),
+            const SizedBox(height: 16),
+            _buildToolsSection(context, healthAuth, colorScheme),
+            const SizedBox(height: 32),
+            
+            // Account Section
+            _buildAccountSection(context, colorScheme),
+            const SizedBox(height: 24),
           ],
-        ));
+        ),
+      ),
+    );
   }
 
   void requestAuthHealt() async {
+    if (!mounted) return;
     try {
       await ref.read(healthProvider).auth();
-      if (context.mounted) context.go('/');
+      if (mounted) {
+        context.go('/');
+      }
     } catch (e) {
-      if (context.mounted) {
+      if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(SnackBar(
           content: Text('$e'),
           behavior: SnackBarBehavior.floating,
@@ -151,4 +105,443 @@ class _SettingsViewState extends ConsumerState<SettingsView> {
       }
     }
   }
+
+  Widget _buildSectionHeader(BuildContext context, String title, ColorScheme colorScheme) {
+    return Text(
+      title,
+      style: Theme.of(context).textTheme.titleLarge?.copyWith(
+        fontWeight: FontWeight.w600,
+        color: colorScheme.onSurface,
+      ),
+    );
+  }
+
+  Widget _buildProfileSection(BuildContext context, ColorScheme colorScheme) {
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.surfaceContainerLow,
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: colorScheme.outline.withOpacity(0.1),
+          width: 1,
+        ),
+      ),
+      child: Material(
+        color: Colors.transparent,
+        child: InkWell(
+          onTap: () => context.push('/setting/profile'),
+          borderRadius: BorderRadius.circular(16),
+          child: Padding(
+            padding: const EdgeInsets.all(20),
+            child: Row(
+              children: [
+                Container(
+                  padding: const EdgeInsets.all(12),
+                  decoration: BoxDecoration(
+                    color: colorScheme.primary.withOpacity(0.1),
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  child: Icon(
+                    Icons.person_outline_rounded,
+                    color: colorScheme.primary,
+                    size: 24,
+                  ),
+                ),
+                const SizedBox(width: 16),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        AppLocalizations.of(context)!.personalData,
+                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                          fontWeight: FontWeight.w600,
+                          color: colorScheme.onSurface,
+                        ),
+                      ),
+                      const SizedBox(height: 4),
+                      Text(
+                        'Manage your profile information',
+                        style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                          color: colorScheme.onSurface.withOpacity(0.7),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+                Icon(
+                  Icons.arrow_forward_ios_rounded,
+                  color: colorScheme.onSurface.withOpacity(0.5),
+                  size: 16,
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildFeaturesSection(BuildContext context, AsyncValue user, ColorScheme colorScheme) {
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.surfaceContainerLow,
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: colorScheme.outline.withOpacity(0.1),
+          width: 1,
+        ),
+      ),
+      child: Column(
+        children: [
+          _buildFeatureToggle(
+            context,
+            colorScheme,
+            icon: Icons.restaurant_outlined,
+            title: AppLocalizations.of(context)!.intermittentFasting,
+            subtitle: 'Track your fasting periods',
+            value: enableFasting,
+            onChanged: (val) {
+              setState(() => enableFasting = !enableFasting);
+              updateProfile(ref,
+                uid: user.value!.uid!,
+                enableFasting: val,
+              );
+            },
+          ),
+          Divider(
+            height: 1,
+            color: colorScheme.outline.withOpacity(0.1),
+            indent: 20,
+            endIndent: 20,
+          ),
+          _buildFeatureToggle(
+            context,
+            colorScheme,
+            icon: Icons.water_drop_outlined,
+            title: AppLocalizations.of(context)!.water,
+            subtitle: 'Monitor your hydration',
+            value: enableWater,
+            onChanged: (val) {
+              setState(() => enableWater = !enableWater);
+              updateProfile(ref,
+                uid: user.value!.uid!,
+                enableWater: val,
+              );
+            },
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildFeatureToggle(
+    BuildContext context,
+    ColorScheme colorScheme, {
+    required IconData icon,
+    required String title,
+    required String subtitle,
+    required bool value,
+    required ValueChanged<bool> onChanged,
+  }) {
+    return Padding(
+      padding: const EdgeInsets.all(20),
+      child: Row(
+        children: [
+          Container(
+            padding: const EdgeInsets.all(12),
+            decoration: BoxDecoration(
+              color: colorScheme.primary.withOpacity(0.1),
+              borderRadius: BorderRadius.circular(12),
+            ),
+            child: Icon(
+              icon,
+              color: colorScheme.primary,
+              size: 24,
+            ),
+          ),
+          const SizedBox(width: 16),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  title,
+                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                    fontWeight: FontWeight.w600,
+                    color: colorScheme.onSurface,
+                  ),
+                ),
+                const SizedBox(height: 4),
+                Text(
+                  subtitle,
+                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                    color: colorScheme.onSurface.withOpacity(0.7),
+                  ),
+                ),
+              ],
+            ),
+          ),
+          Switch(
+            value: value,
+            onChanged: onChanged,
+            activeColor: colorScheme.primary,
+            inactiveThumbColor: colorScheme.outline,
+            inactiveTrackColor: colorScheme.outline.withOpacity(0.3),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildToolsSection(BuildContext context, AsyncValue healthAuth, ColorScheme colorScheme) {
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.surfaceContainerLow,
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: colorScheme.outline.withOpacity(0.1),
+          width: 1,
+        ),
+      ),
+      child: Column(
+        children: [
+          _buildToolItem(
+            context,
+            colorScheme,
+            icon: Icons.devices_outlined,
+            title: AppLocalizations.of(context)!.devices,
+            subtitle: 'Manage connected devices',
+            onTap: () => context.push('/setting/add-device'),
+          ),
+          Divider(
+            height: 1,
+            color: colorScheme.outline.withOpacity(0.1),
+            indent: 20,
+            endIndent: 20,
+          ),
+          _buildHealthItem(context, healthAuth, colorScheme),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildToolItem(
+    BuildContext context,
+    ColorScheme colorScheme, {
+    required IconData icon,
+    required String title,
+    required String subtitle,
+    required VoidCallback onTap,
+  }) {
+    return Material(
+      color: Colors.transparent,
+      child: InkWell(
+        onTap: onTap,
+        borderRadius: BorderRadius.circular(16),
+        child: Padding(
+          padding: const EdgeInsets.all(20),
+          child: Row(
+            children: [
+              Container(
+                padding: const EdgeInsets.all(12),
+                decoration: BoxDecoration(
+                  color: colorScheme.primary.withOpacity(0.1),
+                  borderRadius: BorderRadius.circular(12),
+                ),
+                child: Icon(
+                  icon,
+                  color: colorScheme.primary,
+                  size: 24,
+                ),
+              ),
+              const SizedBox(width: 16),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      title,
+                      style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                        fontWeight: FontWeight.w600,
+                        color: colorScheme.onSurface,
+                      ),
+                    ),
+                    const SizedBox(height: 4),
+                    Text(
+                      subtitle,
+                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                        color: colorScheme.onSurface.withOpacity(0.7),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              Icon(
+                Icons.arrow_forward_ios_rounded,
+                color: colorScheme.onSurface.withOpacity(0.5),
+                size: 16,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildHealthItem(BuildContext context, AsyncValue healthAuth, ColorScheme colorScheme) {
+    return Padding(
+      padding: const EdgeInsets.all(20),
+      child: Row(
+        children: [
+          Container(
+            padding: const EdgeInsets.all(12),
+            decoration: BoxDecoration(
+              color: colorScheme.primary.withOpacity(0.1),
+              borderRadius: BorderRadius.circular(12),
+            ),
+            child: Icon(
+              Icons.favorite_outline_rounded,
+              color: colorScheme.primary,
+              size: 24,
+            ),
+          ),
+          const SizedBox(width: 16),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  AppLocalizations.of(context)!.health,
+                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                    fontWeight: FontWeight.w600,
+                    color: colorScheme.onSurface,
+                  ),
+                ),
+                const SizedBox(height: 4),
+                Text(
+                  'Connect to Apple Health',
+                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                    color: colorScheme.onSurface.withOpacity(0.7),
+                  ),
+                ),
+              ],
+            ),
+          ),
+          healthAuth.value == true
+              ? Container(
+                  padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
+                  decoration: BoxDecoration(
+                    color: Colors.green.withOpacity(0.1),
+                    borderRadius: BorderRadius.circular(20),
+                  ),
+                  child: Text(
+                    AppLocalizations.of(context)!.connected,
+                    style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                      color: Colors.green,
+                      fontWeight: FontWeight.w600,
+                    ),
+                  ),
+                )
+              : Container(
+                  decoration: BoxDecoration(
+                    gradient: LinearGradient(
+                      colors: [colorScheme.primary, colorScheme.primary.withOpacity(0.8)],
+                      begin: Alignment.topLeft,
+                      end: Alignment.bottomRight,
+                    ),
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  child: Material(
+                    color: Colors.transparent,
+                    child: InkWell(
+                      onTap: requestAuthHealt,
+                      borderRadius: BorderRadius.circular(12),
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                        child: Text(
+                          AppLocalizations.of(context)!.requestAuth,
+                          style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                            color: colorScheme.onPrimary,
+                            fontWeight: FontWeight.w600,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildAccountSection(BuildContext context, ColorScheme colorScheme) {
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.errorContainer.withOpacity(0.1),
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: colorScheme.error.withOpacity(0.2),
+          width: 1,
+        ),
+      ),
+      child: Material(
+        color: Colors.transparent,
+        child: InkWell(
+          onTap: () async {
+            await ref.read(logoutUserProvider.future);
+          },
+          borderRadius: BorderRadius.circular(16),
+          child: Padding(
+            padding: const EdgeInsets.all(20),
+            child: Row(
+              children: [
+                Container(
+                  padding: const EdgeInsets.all(12),
+                  decoration: BoxDecoration(
+                    color: colorScheme.error.withOpacity(0.1),
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  child: Icon(
+                    Icons.logout_rounded,
+                    color: colorScheme.error,
+                    size: 24,
+                  ),
+                ),
+                const SizedBox(width: 16),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        AppLocalizations.of(context)!.logout,
+                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                          fontWeight: FontWeight.w600,
+                          color: colorScheme.error,
+                        ),
+                      ),
+                      const SizedBox(height: 4),
+                      Text(
+                        'Sign out of your account',
+                        style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                          color: colorScheme.error.withOpacity(0.7),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+                Icon(
+                  Icons.arrow_forward_ios_rounded,
+                  color: colorScheme.error.withOpacity(0.5),
+                  size: 16,
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+
 }
