@@ -1,11 +1,11 @@
-String? weightRequired(value) {
+String? weightRequired(String? value) {
   if (value == null || value.isEmpty) {
     return 'Weight Required';
   }
   return null;
 }
 
-String? heightRequired(value) {
+String? heightRequired(String? value) {
   if (value == null || value.isEmpty) {
     return 'Weight Required';
   }
