@@ -3,7 +3,25 @@ import 'package:flutter/material.dart';
 class DynamicDialog extends StatefulWidget {
   final String? title;
   final String? body;
-  const DynamicDialog({super.key, this.title, this.body});
+  final String? primaryButtonText;
+  final String? secondaryButtonText;
+  final VoidCallback? onPrimaryPressed;
+  final VoidCallback? onSecondaryPressed;
+  final IconData? icon;
+  final bool showCloseButton;
+  
+  const DynamicDialog({
+    super.key,
+    this.title,
+    this.body,
+    this.primaryButtonText,
+    this.secondaryButtonText,
+    this.onPrimaryPressed,
+    this.onSecondaryPressed,
+    this.icon,
+    this.showCloseButton = true,
+  });
+  
   @override
   DynamicDialogState createState() => DynamicDialogState();
 }
@@ -11,19 +29,168 @@ class DynamicDialog extends StatefulWidget {
 class DynamicDialogState extends State<DynamicDialog> {
   @override
   Widget build(BuildContext context) {
-    // You can change the UI as per
-    // your requirement or choice
-    return AlertDialog(
-      title: Text(widget.title ?? ''),
-      actions: <Widget>[
-        OutlinedButton.icon(
-            label: const Text('Close'),
-            onPressed: () {
-              Navigator.pop(context);
-            },
-            icon: const Icon(Icons.close))
-      ],
-      content: Text(widget.body ?? ''),
+    final colorScheme = Theme.of(context).colorScheme;
+    final theme = Theme.of(context);
+    
+    return Dialog(
+      backgroundColor: Colors.transparent,
+      child: Container(
+        constraints: const BoxConstraints(maxWidth: 400),
+        decoration: BoxDecoration(
+          color: colorScheme.surface,
+          borderRadius: BorderRadius.circular(24),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.1),
+              blurRadius: 20,
+              offset: const Offset(0, 8),
+            ),
+          ],
+        ),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            // Header with icon and title
+            if (widget.icon != null || widget.title != null)
+              Container(
+                padding: const EdgeInsets.all(24),
+                child: Column(
+                  children: [
+                    if (widget.icon != null) ...[
+                      Container(
+                        padding: const EdgeInsets.all(16),
+                        decoration: BoxDecoration(
+                          color: colorScheme.primaryContainer,
+                          borderRadius: BorderRadius.circular(16),
+                        ),
+                        child: Icon(
+                          widget.icon!,
+                          color: colorScheme.onPrimaryContainer,
+                          size: 32,
+                        ),
+                      ),
+                      const SizedBox(height: 16),
+                    ],
+                    if (widget.title != null)
+                      Text(
+                        widget.title!,
+                        style: theme.textTheme.headlineSmall?.copyWith(
+                          fontWeight: FontWeight.w700,
+                          color: colorScheme.onSurface,
+                        ),
+                        textAlign: TextAlign.center,
+                      ),
+                  ],
+                ),
+              ),
+            
+            // Body content
+            if (widget.body != null)
+              Padding(
+                padding: EdgeInsets.fromLTRB(
+                  24,
+                  widget.title != null || widget.icon != null ? 0 : 24,
+                  24,
+                  24,
+                ),
+                child: Text(
+                  widget.body!,
+                  style: theme.textTheme.bodyLarge?.copyWith(
+                    color: colorScheme.onSurfaceVariant,
+                    height: 1.5,
+                  ),
+                  textAlign: TextAlign.center,
+                ),
+              ),
+            
+            // Action buttons
+            Padding(
+              padding: const EdgeInsets.fromLTRB(24, 0, 24, 24),
+              child: Row(
+                children: [
+                  // Secondary button
+                  if (widget.secondaryButtonText != null || widget.showCloseButton)
+                    Expanded(
+                      child: Container(
+                        height: 48,
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(12),
+                          border: Border.all(
+                            color: colorScheme.outline.withOpacity(0.3),
+                            width: 1,
+                          ),
+                        ),
+                        child: Material(
+                          color: Colors.transparent,
+                          child: InkWell(
+                            borderRadius: BorderRadius.circular(12),
+                            onTap: widget.onSecondaryPressed ?? () {
+                              Navigator.of(context).pop();
+                            },
+                            child: Center(
+                              child: Text(
+                                widget.secondaryButtonText ?? 'Close',
+                                style: theme.textTheme.labelLarge?.copyWith(
+                                  fontWeight: FontWeight.w600,
+                                  color: colorScheme.onSurfaceVariant,
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                  
+                  // Spacing between buttons
+                  if ((widget.secondaryButtonText != null || widget.showCloseButton) && 
+                      widget.primaryButtonText != null)
+                    const SizedBox(width: 12),
+                  
+                  // Primary button
+                  if (widget.primaryButtonText != null)
+                    Expanded(
+                      child: Container(
+                        height: 48,
+                        decoration: BoxDecoration(
+                          gradient: LinearGradient(
+                            colors: [
+                              colorScheme.primary,
+                              colorScheme.primary.withOpacity(0.8),
+                            ],
+                          ),
+                          borderRadius: BorderRadius.circular(12),
+                          boxShadow: [
+                            BoxShadow(
+                              color: colorScheme.primary.withOpacity(0.3),
+                              blurRadius: 8,
+                              offset: const Offset(0, 2),
+                            ),
+                          ],
+                        ),
+                        child: Material(
+                          color: Colors.transparent,
+                          child: InkWell(
+                            borderRadius: BorderRadius.circular(12),
+                            onTap: widget.onPrimaryPressed,
+                            child: Center(
+                              child: Text(
+                                widget.primaryButtonText!,
+                                style: theme.textTheme.labelLarge?.copyWith(
+                                  fontWeight: FontWeight.w600,
+                                  color: Colors.white,
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
     );
   }
 }
